Private use codepoints should not be counted as 'control characters'
This is a non-backwards compatible change. It only affects utf8.match,
specifically when '%c' is used. Previously, %c (which is for matching
'control characters') would match private use codepoints, such as those
above U+F0000.

This is very non-intuitive and I doubt that any user of the library
would expect this. (Though, Hyrum's Law...)

Adjust the data tables so that utf8.match does not behave this way.